tweak keyboard input state type

+++ Include/PlatformGameInputHandlingAPI.h
@@ -76,8 +76,8 @@ enum : uint32
 
 typedef struct keycode_state
 {
-    uint32 isDown = false; // is the key currently down
-    uint32 numStateChanges = 0; // number of times the key went up/down
+    uint8 isDown = false; // is the key currently down
+    uint8 numStateChanges = 0; // number of times the key went up/down
 } KeycodeState;
 
 typedef struct mousecode_state
